% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLEV-bivariate.r
\name{BSMGL-EV}
\alias{BSMGL-EV}
\alias{pcMGLEV180.bivar}
\alias{dcMGLEV180.bivar}
\title{Bivarite surivival MGL-EV copula}
\usage{
pcMGLEV180.bivar(u1, u2, param)

dcMGLEV180.bivar(u1, u2, param)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \eqn{\left\[0,1\right\]}.}

\item{param}{copula parameter, denoted by delta}
}
\description{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\examples{
pcMGLEV180.bivar(u1 = c(0.3, 0.9), u2 = c(0.5, 0.8), param = 2)
dcMGLEV180.bivar(u1 = 0.5, u2 = 0.78, param = 1.8)
}
