% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL-bivariate.r
\name{BSMGL}
\alias{BSMGL}
\alias{dcMGL180.bivar}
\alias{pcMGL180.bivar}
\alias{rcMGL180.bivar}
\title{Bivarite Survival MGL copula}
\usage{
dcMGL180.bivar(u1, u2, pars)

pcMGL180.bivar(u1, u2, pars)

rcMGL180.bivar(n, pars)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \eqn{\left\[0,1\right\]}.}

\item{pars}{copula parameter, denoted by delta}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\examples{
dcMGL180.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
pcMGL180.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
rcMGL180.bivar(n = 200, pars = 0.8)
}
