% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL.bivariate.r
\name{dcGLMGA.multi}
\alias{dcGLMGA.multi}
\title{Bivarite MGL copula}
\usage{
dcGLMGA.multi(u, pars, log = FALSE)
}
\arguments{
\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{u_1, u_2}{numeric vectors of equal length with values in \link{0,1}.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
Bivarite MGL copula
}
\examples{
dcMGL.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
pcMGL.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
}
