% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGB2.bivariate.r
\name{dcMGB2.bivar}
\alias{dcMGB2.bivar}
\title{Bivarite MGL copula}
\usage{
dcMGB2.bivar(u1, u2, pars1, pars2, pars3)
}
\arguments{
\item{pars1, pars2, pars3}{copula parameters, denoted by p_1, p_2, q}

\item{u_1, u_2}{numeric vectors of equal length with values in \link{0,1}.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
Density (dCopula), distribution function (pCopula), and random generation (rCopula) for a copula object.
}
\examples{
dcMGB2.bivar(u1 = c(0, 0.5), u2 = c(0.9, 0), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)
pcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)

}
