% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL.multi.r
\name{dcMGL.multi}
\alias{dcMGL.multi}
\title{d-dimensional MGL copula}
\usage{
dcMGL.multi(u, pars, log = FALSE)
}
\arguments{
\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{d}{d-dimensions}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
Density (dCopula), distribution function (pCopula), and random generation (rCopula) for a copula object.
}
\examples{
dcMGL.multi(u = cbind(c(0.6, 0.1, 0.5), c(0.3, 0.9, 0.2)), pars = 10, log = FALSE)
dcMGL.multi(u = cbind(c(0.6, 0.1), c(0.3, 0.9), c(0.5, 0.6)), pars = 2, log = T)
pcMGL.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)

Usim <- rcMGL.multi(n = 1000, d = 2, param = 1)
plot(Usim)
}
