% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLEV.bivariate.r
\name{dcMGLEV.bivar}
\alias{dcMGLEV.bivar}
\title{Bivarite MGL-EV copula}
\usage{
dcMGLEV.bivar(u1, u2, param)
}
\arguments{
\item{param}{copula parameter, denoted by delta}

\item{u_1, u_2}{numeric vectors of equal length with values in \link{0,1}.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
Bivarite MGL-EV copula
}
\examples{
dcMGLEV.bivar(u1 = 0.001, u2 = 0.999, param = 1)
pcMGLEV.bivar(u1 = c(0.3, 0.9), u2 = c(0.5, 0.8), param = 2)
}
