% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGA-distribution.r, R/MGLMGA-distribution.r
\name{dGLMGA}
\alias{dGLMGA}
\title{The GLMGA distribution}
\usage{
dGLMGA(y, sigma, a, b, log = FALSE)

dGLMGA(y, sigma, a, b, log = FALSE)
}
\arguments{
\item{y, q}{vector of quantiles.}

\item{sigma}{parameter}

\item{a}{parameter}

\item{b}{parameter}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{p}{vector of probabilities.}
}
\value{
Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.

Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.
}
\description{
The GLMGA distribution

The GLMGA distribution
}
\examples{
# density function at value 0.5 and 0.1
dLMGA(c(0.5,0.1), sigma = 2, a = 2, b = 3, log = FALSE)

# density function at value 0.5 and 0.1
dLMGA(c(0.5,0.1), sigma = 2, a = 2, b = 3, log = FALSE)

}
