% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLMGA-distribution.r
\name{dMGLMGA}
\alias{dMGLMGA}
\title{The bivariate GLMGA distribution}
\usage{
dMGLMGA(y1, y2, sigma, a, b)
}
\arguments{
\item{y1}{vector of quantiles.}

\item{y2}{vector of quantiles.}

\item{sigma}{vector of parameters with length = 2.}

\item{a}{common parameter with length = 1.}

\item{b}{vector of parameters with length = 2.}
}
\value{
Density, for the bivariate GLMGA distribution with parameters sigma, a and b.
}
\description{
The bivariate GLMGA distribution
}
\examples{
n.grid <- 20
sigma <- c(0.5, 0.5)
a <- 20
b <- c(5, 5)
xgrid <- ygrid <- seq(0.01, 20, length.out = n.grid)
grid <- expand.grid("u1" = xgrid, "u2" = ygrid)
mtrx3d <- matrix(0, nrow = nrow(grid), ncol = 3)
mtrx3d[, 1] <- grid[, 1]
mtrx3d[, 2] <- grid[, 2]
for (i in 1:nrow(mtrx3d)) {
  mtrx3d[i, 3] <- dMGLMGA(
    y1 = grid[i, 1], y2 = grid[i, 2],
    sigma = sigma,
    a = a, b = b
  )
}
head(mtrx3d)
}
