% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error-copua.r
\name{error.copula}
\alias{error.copula}
\title{the squared fit error}
\usage{
error.copula(
  XY,
  copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "Normal", "MGB2", "t"),
  lowerLimit,
  upperLimit,
  pars,
  ...
)
}
\arguments{
\item{XY}{the 2-dim loss data}

\item{copula}{copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "Normal", "MGB2", "t"}

\item{lowerLimit}{The lower limit of integration, a vector for hypercubes.}

\item{upperLimit}{The upper limit of integration, a vector for hypercubes.}

\item{pars}{copula parameters}

\item{...}{All other arguments passed to the function f.}
}
\value{
vector of length nComp; the integral of integrand over the hypercube
}
\description{
the squared fit error
}
