% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL-bivariate.r
\name{hBMGL}
\alias{hBMGL}
\alias{hcMGL.bivar}
\alias{hcMGL180.bivar}
\title{Conditional Distribution Function of a Bivariate Copula}
\usage{
hcMGL.bivar(u1, u2, pars)

hcMGL180.bivar(u1, u2, pars)
}
\arguments{
\item{u1}{numeric vectors of equal length with values in \eqn{[0,1]}.}

\item{u2}{numeric vectors of equal length with values in \eqn{[0,1]}.}

\item{pars}{numeric; single number or vector of size length(u1); copula parameter > 0.}
}
\value{
hcMGL.bivar returns a list with
\itemize{
\item hfunc1: \eqn{\partial C(u_1,u_2) / \partial u_1,}
\item hfunc2: \eqn{\partial C(u_1,u_2) / \partial u_2,}
}
}
\description{
Conditional Distribution Function of a Bivariate Copula
}
\details{
The h-function is defined as the conditional distribution function of a bivariate copula, i.e.,
\deqn{h_1(u_2|u_1,\delta) := P(U_2 \leq u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,}

\deqn{h_2(u_1|u_2,\delta) := P(U_1 \leq u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,}

where \eqn{(U_1, U_2) \sim C}, and \eqn{C} is a bivariate copula distribution function with parameter(s) \eqn{\delta}. For more details see Aas et al. (2009).
}
\examples{
hcMGL.bivar(u1 = c(0.1, 0.001, 0.3), u2 = c(0, 0.9999, 0.88), pars = 2)
hcMGL180.bivar(u1 = c(0.1, 0.001, 0.3), u2 = c(0, 0.9999, 0.88), pars = 2)
}
\references{
Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009). Pair-copula constructions of multiple dependence. Insurance: Mathematics and Economics 44 (2), 182-198.
}
