% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGB2.bivariate.r
\name{hcMGB2.bivar}
\alias{hcMGB2.bivar}
\title{Conditional Distribution Function of a Bivariate Copula}
\usage{
hcMGB2.bivar(u1, u2, pars1, pars2, pars3)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \link{0,1}.}

\item{pars1, pars2, pars3}{copula parameters, denoted by p_1, p_2, q}
}
\description{
Conditional Distribution Function of a Bivariate Copula
}
\details{
The h-function is defined as the conditional distribution function of a bivariate copula, i.e.,
h_1(u_2|u_1,θ) := P(U_2 ≤ u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,

h_2(u_1|u_2,θ) := P(U_1 ≤ u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,

where (U_1, U_2) \sim C, and C is a bivariate copula distribution function with parameter(s) θ. For more details see Aas et al. (2009).
}
\examples{
hMGB2.bivar(u1 = c(0.3, 0.4), u2 = c(0.4, 0.5), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)

}
