% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL.bivariate.r
\name{hcMGL.bivar}
\alias{hcMGL.bivar}
\title{Conditional Distribution Function of a Bivariate Copula}
\usage{
hcMGL.bivar(u1, u2, pars)
}
\arguments{
\item{u1}{numeric vectors of equal length with values in \link{0,1}.}

\item{u2}{numeric vectors of equal length with values in \link{0,1}.}

\item{pars}{numeric; single number or vector of size length(u1); copula parameter > 0.}
}
\description{
Conditional Distribution Function of a Bivariate Copula
}
\details{
The h-function is defined as the conditional distribution function of a bivariate copula, i.e.,
h_1(u_2|u_1,θ) := P(U_2 ≤ u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,

h_2(u_1|u_2,θ) := P(U_1 ≤ u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,

where (U_1, U_2) \sim C, and C is a bivariate copula distribution function with parameter(s) θ. For more details see Aas et al. (2009).
}
\examples{
hcMGL.bivar(u1 = c(0.1, 0.001, 0.3), u2 = c(0, 0.9999, 0.88), pars = 2)
}
