% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGB2-bivariate.r
\name{MGB2}
\alias{MGB2}
\alias{dcMGB2.bivar}
\alias{pcMGB2.bivar}
\alias{hcMGB2.bivar}
\title{Bivarite MGB2 copula}
\usage{
dcMGB2.bivar(u1, u2, pars1, pars2, pars3)

pcMGB2.bivar(u1, u2, pars1, pars2, pars3)

hcMGB2.bivar(u1, u2, pars1, pars2, pars3)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \eqn{\left\[0,1\right\]}.}

\item{pars1, pars2, pars3}{copula parameters, denoted by p_1, p_2, q}
}
\description{
Density, distribution function, and random generation for the 2-dimensional MGB2.
}
\examples{
dcMGB2.bivar(u1 = c(0.2, 0.5), u2 = c(0.9, 0.2), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)
pcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)
}
