% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL-mle-mixed.r
\name{MGL.mle.mixed}
\alias{MGL.mle.mixed}
\title{Fitting bivariate MGL copula models}
\usage{
MGL.mle.mixed(
  obs,
  U,
  U_,
  f,
  copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "Normal", "MGB2", "t"),
  umin,
  hessian = TRUE,
  initpar,
  ...
)
}
\arguments{
\item{obs}{two-dimenstional matrix for observations}

\item{U}{two-dimenstional matrix with values in \eqn{\left\[0,1\right\]}.}

\item{U_}{asd}

\item{f}{values of the density function for marginal distribution}

\item{copula}{copula 'MGL', 'MGL180', "MGL-EV", "MGL-EV180", "MGB2", "Normal" , "Student-t"}

\item{umin}{sd}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{additional arguments to be passed to f.}
}
\value{
A list containing the following components
}
\description{
MGL.reg is used to fit bivariate MGL copula regression models.
}
