% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL-mle.r
\name{MGL.mle}
\alias{MGL.mle}
\title{Fitting bivariate MGL copula models}
\usage{
MGL.mle(
  U,
  copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "Normal", "MGB2", "t"),
  hessian = TRUE,
  initpar,
  ...
)
}
\arguments{
\item{U}{two-dimenstional matrix with values in \eqn{[0,1]}.}

\item{copula}{copula 'MGL', 'MGL180', "MGL-EV", "MGL-EV180", "MGB2", "Normal" , "Student-t".}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{additional arguments, see \code{\link[stats]{nlm}} for more details.}
}
\value{
A list containing the following components:
\itemize{
\item loglike: the value of the estimated maximum of the loglikelihood function.
\item copula: the name of the fitted copula. "MGL180" and "MGL-EV180" denote the survival MGL and MGL-EV copula respectively.
\item estimates: the point at which the maximum value of the loglikelihood is obtained.
\item se: the standard errors of the estimators.
\item AIC, BIC: the goodness fit of the regression models.
\item hessian: the hessian at the estimated maximum of the loglikelihood (if requested).
}
}
\description{
\code{MGL.mle} is used to fit bivariate copula regression models via maximum likelihood (ML) method for two continuous variables.
}
\details{
\itemize{
\item copula: "MGL180" and "MGLEV180" denote the survival MGL and survival MGL-EV copula respectively.
}
}
