% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLReg-multi.r
\name{MGL.reg.multi}
\alias{MGL.reg.multi}
\title{Fitting d-dimensional MGL and survival MGL regression models}
\usage{
MGL.reg.multi(U, X, copula = c("MGL", "MGL180"), hessian = TRUE, initpar, ...)
}
\arguments{
\item{U}{d-dimenstional matrix with values in \eqn{[0,1]}.}

\item{X}{design matrix.}

\item{copula}{'MGL', 'MGL180'.}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{additional arguments, see \code{\link[stats]{optim}} for more details.}
}
\value{
A list containing the following components:
\itemize{
\item loglike: the value of the estimated maximum of the loglikelihood function.
\item copula: the name of the fitted copula. "MGL180" and "MGL-EV180" denote the survival MGL and MGL-EV copula respectively.
\item estimates: the point at which the maximum value of the loglikelihood is obtained.
\item se: the standard errors of the estimators.
\item AIC, BIC: the goodness fit of the regression models.
\item hessian: the hessian at the estimated maximum of the loglikelihood (if requested).
}
}
\description{
\code{MGL.reg.multi} is used to fit d-dimensional MGL and survival MGL regression models.
}
\details{
\itemize{
\item copula: "MGL180" denotes the survival MGL and survival MGL-EV copula respectively.
\item For "MGL" and "MGL180" regression model, the copula parameter \deqn{\delta_i = \exp(X\beta),} where \eqn{\beta} is the vector of coefficients to be estimated in the copula regression.
}
}
