% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLReg.r
\name{MGL.reg}
\alias{MGL.reg}
\title{Fitting bivariate/multivariate MGL and MGL-EV copula regression models.}
\usage{
MGL.reg(
  U,
  X,
  copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "MGB2"),
  hessian = TRUE,
  initpar,
  ...
)
}
\arguments{
\item{U}{two-dimensional matrix with values in \eqn{[0,1]}.}

\item{X}{design matrix}

\item{copula}{'MGL', 'MGL180', "MGL-EV", "MGL-EV180", "Gumbel", "MGB2".}

\item{hessian}{Logical. Should a numerically differentiated Hessian matrix be returned?}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{additional arguments, see \code{\link[stats]{optim}} for more details.}
}
\value{
A list containing the following components:
\itemize{
\item loglike: the value of the estimated maximum of the loglikelihood function.
\item copula: the name of the fitted copula. "MGL180" and "MGL-EV180" denote the survival MGL and MGL-EV copula respectively.
\item estimates: the point at which the maximum value of the loglikelihood is obtained.
\item se: the standard errors of the estimators.
\item AIC, BIC: the goodness fit of the regression models.
\item hessian: the hessian at the estimated maximum of the loglikelihood (if requested).
}
}
\description{
\code{MGL.reg} is used to fit bivariate MGL and MGL-EV copula regression models for two continuous variables.
}
\details{
The estimation method is performed via \code{\link[stats]{optim}} function. Y1 and Y2 are both continuous variables.

copula: "MGL180" and "MGLEV180" denote the survival MGL and survival MGL-EV copula respectively.
\itemize{
\item For "Gumbel" regression model, the copula parameter \deqn{\delta_i = \exp(X_i^T\beta) + 1,} where \eqn{{X}_{i}=(1,x_{i1}...,x_{ik})} denotes the vector of covariates and \eqn{\beta} is the vector of coefficients to be estimated in the copula regression.
\item For "MGL", "MGL180", "MGL-EV", "MGL-EV180" regression model, the copula parameter \deqn{\delta_i = \exp(X_i^T\beta).}
\item For "MGB2", the copula parameter \deqn{\q_i = \exp(X_i^T\beta)} and \deqn{(p_1,p_2)} remain to be constant.
Note that the regression modelling can be extended to the high-dimensional case when copula is "MGL180" and "MGL".
}
}
\examples{
# 10-dimensional regression models
set.seed(111)
d <- 10
n <- 1000 # sample size
beta.true <- c(-0.6, 0.5, 0.2) # true regression coefficients
x1 <- rnorm(n, 0, 1)
x2 <- rnorm(n, 0, 1)
X <- model.matrix(~ x1 + x2) # design matrix
delta.sim <- as.vector(exp(X\%*\%beta.true)) # true copula parameters
Usim <- matrix(0, nrow = n, ncol = d)
for (i in 1:n){
  Usim[i, ] <- rcMGL.multi(n = 1, d = d, pars = delta.sim[i])
}
m.MGLMGA <- MGL.reg(U = Usim, copula = "MGL",
                       X = X, method = "Nelder-Mead",
                       initpar = c(-0.32, 0.001, 0.001)
)
m.MGLMGA

}
