% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMGL.r
\name{pcGLMGA.multi}
\alias{pcGLMGA.multi}
\title{d-dimensional MGL copula}
\usage{
pcGLMGA.multi(u, pars)
}
\arguments{
\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\description{
d-dimensional MGL copula
}
\examples{
pcGLMGA.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)



}
