% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGA-distribution.r, R/MGLMGA-distribution.r
\name{pGLMGA}
\alias{pGLMGA}
\title{The GLMGA distribution}
\usage{
pGLMGA(y, sigma, a, b)

pGLMGA(y, sigma, a, b)
}
\arguments{
\item{y, q}{vector of quantiles.}

\item{sigma}{parameter}

\item{a}{parameter}

\item{b}{parameter}

\item{p}{vector of probabilities.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\value{
Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.

Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.
}
\description{
The GLMGA distribution

The GLMGA distribution
}
\examples{
# cdf at value 10 and 20.
pLMGA(c(10,20), sigma = 2, a = 2, b = 3)
# cdf at value 10 and 20.
pLMGA(c(10,20), sigma = 2, a = 2, b = 3)
}
