% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGA.r
\name{qGLMGA}
\alias{qGLMGA}
\title{The GLMGA distribution}
\usage{
qGLMGA(u, sigma, a, b)
}
\arguments{
\item{sigma}{parameter}

\item{a}{parameter}

\item{b}{parameter}

\item{y, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\value{
Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.
}
\description{
The GLMGA distribution
}
\examples{
# quantile function at level 50\% and 10\%
qLMGA(c(0.5,0.1), sigma = 2, a = 2, b = 3)
}
