% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMGL.r
\name{rcGLMGA.multi}
\alias{rcGLMGA.multi}
\title{d-dimensional MGL copula}
\usage{
rcGLMGA.multi(n, d, param)
}
\arguments{
\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\description{
d-dimensional MGL copula
}
\examples{
Usim <- rcGLMGA.multi(n = 1000, d = 2, param = 1)
Usim

}
