% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL.multi.r
\name{rcMGL.multi}
\alias{rcMGL.multi}
\title{d-dimensional MGL copula}
\usage{
rcMGL.multi(n, d, param)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{d}{d-dimensions}

\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
d-dimensional MGL copula
}
\examples{
dcMGL.multi(u = cbind(c(0.6, 0.1, 0.5), c(0.3, 0.9, 0.2)), pars = 10, log = FALSE)
dcMGL.multi(u = cbind(c(0.6, 0.1), c(0.3, 0.9), c(0.5, 0.6)), pars = 2, log = T)
pcMGL.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)

Usim <- rcMGL.multi(n = 1000, d = 2, param = 1)
plot(Usim)
}
