% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL180.bivariate.r
\name{rcMGL180.bivar}
\alias{rcMGL180.bivar}
\title{Survival bivarite MGL copula}
\usage{
rcMGL180.bivar(n, param)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{u_1, u_2}{numeric vectors of equal length with values in \link{0,1}.}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\value{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\description{
Survival bivarite MGL copula
}
\examples{
dcMGL180.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
pcMGL180.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
Usim <- rcMGL180.bivar(5000, param = 7)
plot(Usim)

}
