% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGA-distribution.r, R/MGLMGA-distribution.r
\name{rGLMGA}
\alias{rGLMGA}
\title{The GLMGA distribution}
\usage{
rGLMGA(n, sigma, a, b)

rGLMGA(n, sigma, a, b)
}
\arguments{
\item{sigma}{parameter}

\item{a}{parameter}

\item{b}{parameter}

\item{y, q}{vector of quantiles.}

\item{p}{vector of probabilities.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}
}
\value{
Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.

Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.
}
\description{
The GLMGA distribution

The GLMGA distribution
}
\examples{
# simulate 10 samples from GLMGA distribution with parameters (2, 2, 3)
rLMGA(n = 10, sigma = 2, a = 2, b = 3)
# simulate 10 samples from GLMGA distribution with parameters (2, 2, 3)
rLMGA(n = 10, sigma = 2, a = 2, b = 3)
}
