% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLEV-bivariate.r
\name{Afunction}
\alias{Afunction}
\title{the Pickands dependence function \eqn{A} in survival MGL-EV copula}
\usage{
Afunction(w, param)
}
\arguments{
\item{w}{\eqn{w \in [0.5,1]}.}

\item{param}{copula parameter in survival MGL-EV copula.}
}
\value{
the value of the Pickands dependence function in \eqn{[0.5,1]}
}
\description{
the Pickands dependence function \eqn{A} in survival MGL-EV copula
}
\details{
\itemize{
\item In the bivariate case
\eqn{d=2} the stable tail dependence function \eqn{\ell} can be represented in terms of the Pickands dependence function \eqn{A}:
}

\deqn{
		\begin{align}
		A_{\delta}\left(w\right)=w I_{{\frac{1}{2}, \frac{1}{\delta}+\frac{1}{2}}}\left[\frac{\left(1-w \right)^{-\delta}}{\left(1-w \right)^{-\delta} + w^{-\delta} }\right]  + \left(1-w\right) I_{{\frac{1}{2}, \frac{1}{\delta}+\frac{1}{2}}}\left[\frac{w^{-\delta}}{\left(1-w \right)^{-\delta} + w^{-\delta} }\right].
		\end{align}
}
where \eqn{A_{\delta}} is the Pickands dependence function of  survival MGL-EV copula.
\itemize{
\item Here \eqn{I_{m,n}^{-1}()} denotes the inverse of the beta cumulative distribution function (or regularized incomplete beta function)
with parameters shape1 = m and shape2 = n
implemented by R's \code{\link[stats]{qbeta}}.
}
}
\examples{
Afunction(w = 0.7, param = 1.2)
}
