% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLEV-bivariate.r
\name{BMGL-EV}
\alias{BMGL-EV}
\alias{pcMGLEV.bivar}
\alias{dcMGLEV.bivar}
\title{Bivarite MGL-EV copula}
\usage{
pcMGLEV.bivar(u1, u2, param)

dcMGLEV.bivar(u1, u2, param)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \eqn{\left\[0,1\right\]}.}

\item{param}{copula parameter, denoted by delta}
}
\description{
Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter delta.
}
\examples{
pcMGLEV.bivar(u1 = c(0.3, 0.9), u2 = c(0.5, 0.8), param = 2)
dcMGLEV.bivar(u1 = 0.001, u2 = 0.999, param = 1)
}
