% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGL-multi.r
\name{cMGL.multi}
\alias{cMGL.multi}
\alias{dcMGL.multi}
\alias{dcMGL180.multi}
\alias{pcMGL.multi}
\alias{pcMGL180.multi}
\alias{rcMGL.multi}
\alias{rcMGL180.multi}
\title{d-dimensional MGL and survival MGL copula}
\usage{
dcMGL.multi(u, pars, log = FALSE)

dcMGL180.multi(u, pars, log = FALSE)

pcMGL.multi(u, pars)

pcMGL180.multi(u, pars)

rcMGL.multi(n, d, pars)

rcMGL180.multi(n, d, pars)
}
\arguments{
\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by \eqn{\delta>0}.}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{d}{d-dimensional}
}
\value{
\itemize{
\item \code{dcMGL.multi}, \code{pcMGL.multi} and \code{rcMGL.multi} gives values of Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter \eqn{\delta>0}.
\item \code{dcMGL180.multi}, \code{pcMGL180.multi} and \code{rcMGL180.multi} gives values of Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter \eqn{\delta>0}.
}
}
\description{
Density, distribution function, and random generation for the MGL and survival MGL copula.
}
\details{
\itemize{
\item The MGL copula density function is given by
}

\eqn{
c^{\text{MGL}}(u_1, \dots, u_d;a)=\frac{\Gamma(a)^{d-1}\Gamma(a+\frac{d}{2})}{\Gamma(a+\frac{1}{2})^d}\frac{\prod_{j=1}^{d}(t(u_j;a)+1)^{a+\frac{1}{2}}}{\left(\sum_{j=1}^{d}t(u_j;a)+1\right)^{a+\frac{d}{2}}},
\quad (u_1, ..., u_d)\in \left[0,1\right]^d,
}

where \eqn{a = \frac{1}{\delta}}, and \eqn{I_{m,n}()}  denotes the beta cumulative distribution function (or regularized incomplete beta function)
with parameters shape1 = m and shape2 = n
implemented by R's \code{\link[stats]{pbeta}}.
\itemize{
\item The MGL copula distribution function has the form of
}

\deqn{
 C^{\text{MGL}}(u_1, \dots, u_d;a) =\mathbb{E}_{\Theta}\left[\prod_{j=1}^{d}\text{erfc}\left(
 \sqrt{t(u_j;a)\Theta}
 \right)\right]
 }
with \eqn{\Theta \sim \text{gamma}(a,1)} and
\deqn{
 t(u_j;a)=\frac{I^{-1}_{\frac{1}{2},{a}}(1-u_j)}{1-I^{-1}_{\frac{1}{2},{a}}(1-u_j)}.
 }
}
\examples{


dcMGL.multi(u = cbind(c(0.6, 0.1, 0.5), c(0.3, 0.9, 0.2)), pars = 2, log = FALSE)

dcMGL.multi(u = cbind(c(0.6, 0.1), c(0.3, 0.9), c(0.5, 0.6)), pars = 2, log = TRUE)
\dontrun{

dcMGL180.multi(u = cbind(c(0.6, 0.1, 0.5), c(0.3, 0.9, 0.2)), pars = 2, log = FALSE)

dcMGL180.multi(u = cbind(c(0.6, 0.1), c(0.3, 0.9), c(0.5, 0.6)), pars = 2, log = TRUE)
}
# 2-dim MGL copula
pcMGL.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)
# 3-dim MGL copula
pcMGL.multi(u = cbind(c(0, 0.2, 0.5), c(0.5, 0.2, 0.5), c(0.01, 0.5, 0.9)), pars = 3)
pcMGL180.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)
Usim <- rcMGL.multi(n = 1000, d = 2, pars = 1)
plot(Usim)
Usim <- rcMGL180.multi(n = 1000, d = 2, pars = 1)
plot(Usim)
}
