% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMGA-distribution.r
\name{GLMGA}
\alias{GLMGA}
\alias{dGLMGA}
\alias{pGLMGA}
\alias{qGLMGA}
\alias{rGLMGA}
\title{The GLMGA distribution}
\usage{
dGLMGA(y, sigma, a, b, log = FALSE)

pGLMGA(y, sigma, a, b)

qGLMGA(u, sigma, a, b)

rGLMGA(n, sigma, a, b)
}
\arguments{
\item{y}{vector of quantiles.}

\item{sigma}{parameter}

\item{a}{parameter}

\item{b}{parameter}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{u}{vector of probabilities.}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}
}
\description{
Density, distribution function, quantile function and random generation for the GLMGA distribution with parameters sigma, a and b.
}
\examples{
# density function at value 0.5 and 0.1
dGLMGA(c(0.5, 0.1), sigma = 2, a = 2, b = 3, log = FALSE)
# cdf at value 10 and 20.
pGLMGA(c(10, 20), sigma = 2, a = 2, b = 3)
# quantile function at level 50\% and 10\%
qGLMGA(c(0.5, 0.1), sigma = 2, a = 2, b = 3)
# simulate 10 samples from GLMGA distribution with parameters (2, 2, 3)
rGLMGA(n = 10, sigma = 2, a = 2, b = 3)
}
