% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGB2-bivariate.r
\name{MGB2}
\alias{MGB2}
\alias{dMGB2}
\alias{pMGB2}
\alias{hMGB2}
\alias{dcMGB2.bivar}
\alias{pcMGB2.bivar}
\alias{hcMGB2.bivar}
\title{Bivarite MGB2 copula}
\usage{
dcMGB2.bivar(u1, u2, pars1, pars2, pars3)

pcMGB2.bivar(u1, u2, pars1, pars2, pars3)

hcMGB2.bivar(u1, u2, pars1, pars2, pars3)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length with values in \eqn{[0,1]}.}

\item{pars1, pars2, pars3}{copula parameters, denoted by \eqn{p_1, p_2, q}.}
}
\value{
\code{dcMGB2.bivar} gives the density

\code{pcMGB2.bivar} gives the distribution function

\code{hcMGB2.bivar} gives the h-functions (hfunc1, hfunc2).
}
\description{
Density, distribution function, and h-functions for the bivariate MGB2 copula proposed in Yang et al.,(2011).
}
\details{
The MGB2 copula with parameters (p1, p2, q) has joint density
\deqn{c(u_1,u_2;p_1,p_2,q)=\frac{\Gamma(q)\Gamma(\sum_{i=1}^2 p_i + q)}{\prod_{i=1}^2\Gamma(p_i+q)}\frac{\prod_{i=1}^2 (1 + x(u_i;p_i,q))^{p_i+q}}{(1 + \sum_{i=1}^{2}x(u_i;p_i,q))^{\sum_{i=1}^{2}p_i+q}},}
for \eqn{p_1, p_2>0, q>0}.
(Here Gamma(a) is the function implemented by R's \code{\link[base]{gamma}} and defined in its help.

The joint cdf of the MGB2 copula is

\deqn{C(u_1,u_2;p_1,p_2,q)=\int_{0}^{+\infty}\prod_{i=1}^{2}G_p(\frac{I_{p_i,q}^{-1}(u_i)}{(1-I_{p_i,q}^{-1}(u_i))\theta})\times \frac{\theta^{-(q+1)}e^{-1/\theta}}{\Gamma(q)}d\theta, }
where \eqn{I_{m,n}^{-1}()} denotes the inverse of the beta cumulative distribution function (or regularized incomplete beta function)
with parameters shape1 = m and shape2 = n
implemented by R's \code{\link[stats]{qbeta}}.

The h-function is defined as the conditional distribution function of a bivariate copula, i.e.,
\deqn{h_1(u_2|u_1,p_1,p_2,q) := P(U_2 \leq u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,}

\deqn{h_2(u_1|u_2,p_1,p_2,q) := P(U_1 \leq u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,}

where \eqn{(U_1, U_2) \sim C}, and \eqn{C} is a bivariate copula distribution function with parameter(s) \eqn{p_1,p_2,q}.
}
\examples{
dcMGB2.bivar(u1 = c(0.2, 0.5), u2 = c(0.9, 0.2), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)
pcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)
hcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)$hfunc1
hcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)$hfunc2
}
\references{
Xipei Yang, Edward W Frees, and Zhengjun Zhang. A generalized beta copula with applications in
modeling multivariate long-tailed data. Insurance: Mathematics and Economics, 49(2):265-284, 2011.
}
