% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLFitting.r
\name{MGL.mle}
\alias{MGL.mle}
\title{Fitting bivariate MGL copula models}
\usage{
MGL.mle(
  U,
  copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180", "Gumbel", "Normal", "MGB2", "t"),
  initpar,
  ...
)
}
\arguments{
\item{U}{two-dimenstional matrix with values in \link{0,1}.}

\item{copula}{'MGL', 'MGL180', "MGL-EV", "MGL-EV180", "MGB2", "Normal" , "Student-t"}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{additional arguments to be passed to f.}

\item{X}{design matrix}
}
\value{
A list containing the following components:
\itemize{
\item minimum   the value of the estimated minimum of f.
\item estimate	  the point at which the minimum value of f is obtained.
\item gradient	the gradient at the estimated minimum of f.
\item hessian	the hessian at the estimated minimum of f (if requested).
}
}
\description{
MGL.reg is used to fit bivariate MGL copula regression models.
}
