% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLReg.r
\name{MGL.reg}
\alias{MGL.reg}
\title{Fitting bivariate MGL copula regression models}
\usage{
MGL.reg(U, X, copula = c("MGL", "MGL180", "MGL-EV", "MGL-EV180"), initpar, ...)
}
\arguments{
\item{U}{two-dimenstional matrix with values in \link{0,1}.}

\item{X}{design matrix}

\item{copula}{'MGL', 'MGL180', "MGL-EV", "MGL-EV180"}

\item{initpar}{Initial values for the parameters to be optimized over.}

\item{...}{Further arguments to be passed to fn and gr in optiom.}
}
\value{
For optim, a list with components:
}
\description{
MGL.reg is used to fit bivariate MGL copula regression models.
}
