% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMGL.r
\name{pcGLMGA.multi}
\alias{pcGLMGA.multi}
\title{d-dimensional MGL copula}
\usage{
pcGLMGA.multi(u, pars)
}
\arguments{
\item{u}{d-dimensional matrix}

\item{pars}{copula parameter, denoted by delta}

\item{log}{logical; if TRUE, probabilities/densities p are returned as log(p).}

\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{d}{d-dimensions}
}
\description{
d-dimensional MGL copula
}
\examples{
dcGLMGA.multi(u = cbind(c(0.6, 0.1, 0.5), c(0.3, 0.9, 0.2)), pars = 10, log = FALSE)
dcGLMGA.multi(u = cbind(c(0.6, 0.1), c(0.3, 0.9), c(0.5, 0.6)), pars = 2, log = T)
pcGLMGA.multi(u = cbind(c(0.5, 0.5), c(0.01, 0.9)), pars = 3)

Usim <- rcGLMGA.multi(n = 1000, d = 2, param = 1)
Usim
}
